requirements.txt
boto3==1.12.39
botocore==1.15.39
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.2.1
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
oauthlib==3.1.0
Pillow==6.2.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1